\name{WebDevelopR-package}
\alias{WebDevelopR-package}
\alias{WebDevelopR}
\docType{package}
\title{
Website development package for R
}
\description{
Two versions of the package are available.  One provides only the basic script required to facilitate communication between a site user's web browser and an R script.  The other consists of a basic web application with a form for the site user to choose options for the analysis and a results page showing the results of the analysis.  The package can be viewed as an extension of CGIwithR, fixing a few technical problems associated with large data files and long computations, addressing some security issues, allowing the web programmer to choose the data interchange format to be used in communications between the server and the browser, and providing a template website to reduce the programming burden on the user.
}
\details{
\tabular{ll}{
Package: \tab WebDevelopR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-25\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Evan Ray, Peter Krafft, John Staudenmayer

Maintainer: Evan Ray <ray@math.umass.edu>
}
\references{
Firth D (2003). "CGIwithR: Facilities for Processing Web Forms Using R." Journal of Statistical Software, 8(10), 1-8. ISSN 1548-7660. URL http://www.jstatsoft.org/v08/i10.
}
\keyword{ package }
